requirements.txt
asgiref
Django
django-filter
djangorestframework
pytz
sqlparse
django-cors-headers